update to ne html model. del and arrows
1. Transition to new html ui framework
2. support for del and arrow keys
3. support to move by one using arrow keys
