+++ index.html
@@ -37,7 +37,8 @@
   </style>
 </head>
 
-<body>
+<body data-codeword="mypypdftools">
+  
     <div id="loader" class="loader" style="display: none;"></div> <!--circle displayed while working-->
     <h2 style="text-align: center;">PDF File Manipulation Tools</h2>
     <h5 style="text-align: center;">All the tools assume use of pdf files</h5>
@@ -126,5 +127,6 @@
 </body>
 
   <script src='uiclient.js'></script>
+  <script src='myfunc.js'></script>
 
 </html>
